requirements.txt
crispy-bootstrap5==2024.2
Django==5.0.5
django-crispy-forms==2.1
django-role-permissions==3.2.0